backend/requirements.txt
coverage==7.5.3
factory_boy==3.3.0
fastapi==0.111.0
pandas==2.2.2
pydantic==2.7.4
pytest==8.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.30
